pyproject.toml
[project]
name = "ai-scheduling-agent"
version = "0.1.0"
package-mode = false

description = "An AI agent for scheduling medical appointments."
authors = [
    {name = "Himanshu Gupta",email = "ghimax23@gmail.com"}
]
readme = "README.md"
requires-python =">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "langchain (>=0.3.27,<0.4.0)",
    "langgraph (>=0.6.6,<0.7.0)",
    "pandas (>=2.3.2,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "streamlit (>=1.49.1,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "langchain-ollama (>=0.3.7,<0.4.0)",
    "langchain-groq (>=0.3.7,<0.4.0)",
    "faker (>=37.6.0,<38.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
